Better error messages and hoping that CI will tell me what is wrong

+++ uspec/test_specs/kill_this_script.sh
@@ -1,13 +1,25 @@
 #!/usr/bin/env sh
 
+command_exists() { command -v "$1" 1>&- 2>&-; }
+
 THISDIR="$(realpath "$(dirname "$0")")"
 ROOTDIR="$(realpath "$THISDIR/../..")"
 
 path="$1"
 name="$(basename "$path")"
 
+if ! [ -f "$path" ]; then
+  echo "file not found: $path"
+  exit 255
+fi
+
 mkdir -v -p tmp
-bundle exec "$ROOTDIR/bin/uspec" "$path" > "tmp/$name.output" 2>&1 &
+if command_exists bundle; then
+  bundle exec "$ROOTDIR/bin/uspec" "$path" > "tmp/$name.output" 2>&1 &
+else
+  "$ROOTDIR/bin/uspec" "$path" > "tmp/$name.output" 2>&1 &
+fi
+
 pid=$!
 sleep 1
 kill $pid
